Split up sdk site types
