+++ site/types/schema/content/career.ts
@@ -1,0 +1,13 @@
+export interface Career {
+	id?: string;
+	status?: string;
+	sort?: number | null;
+	title?: string;
+	summary?: string;
+	body?: string | null;
+	date_created?: string | null;
+	category?: string;
+	employment?: string;
+	slug?: string;
+	apply?: string | null;
+}
