+++ site/types/schema/content/case-study.ts
@@ -1,0 +1,3 @@
+export interface CaseStudy {
+	id?: string;
+}
