+++ site/types/schema/content/event-info.ts
@@ -1,0 +1,3 @@
+export interface EventInfo {
+	[key: string]: any;
+}
