+++ site/types/schema/content/index.ts
@@ -1,0 +1,14 @@
+export type * from './article.js';
+export type * from './blog.js';
+export type * from './career.js';
+export type * from './case-study.js';
+export type * from './event-info.js';
+export type * from './form.js';
+export type * from './home-release-tile.js';
+export type * from './page/index.js';
+export type * from './report.js';
+export type * from './resource.js';
+export type * from './site-banner.js';
+export type * from './testimonial.js';
+export type * from './user-group.js';
+export type * from './video.js';
