+++ site/types/schema/content/page/block/block-feature-grid.ts
@@ -1,0 +1,4 @@
+export interface BlockFeaturegrid {
+	id?: string;
+	features?: { [key: string]: any } | null;
+}
