+++ site/types/schema/content/page/block/block-hero-form.ts
@@ -1,0 +1,8 @@
+import type { Form } from '../../form.js';
+
+export interface BlockHeroForm {
+	id?: string;
+	form?: (string | Form) | null;
+	heading?: string | null;
+	subheading?: string | null;
+}
