+++ site/types/schema/content/page/block/block-hero-headline.ts
@@ -1,0 +1,9 @@
+import type { ComponentButtonGroup } from '../component/index.js';
+
+export interface BlockHeroHeadline {
+	id?: string;
+	/** Use italics to make text gradient. */
+	heading?: string | null;
+	subheading?: string | null;
+	button_group?: (string | ComponentButtonGroup) | null;
+}
