+++ site/types/schema/content/page/block/block-hero-rotator.ts
@@ -1,0 +1,3 @@
+export interface BlockHeroRotator {
+	id?: string;
+}
