+++ site/types/schema/content/page/block/block-page-header.ts
@@ -1,0 +1,12 @@
+import { ComponentButtonGroup } from '../component/index.js';
+
+export interface BlockPageheader {
+	id?: string;
+	heading?: string | null;
+	preheading?: string | null;
+	alignment?: string | null;
+	subheading?: string | null;
+	buttons?: (string | ComponentButtonGroup) | null;
+	block_content?: string;
+	block_styling?: string;
+}
