+++ site/types/schema/content/page/block/block-separator.ts
@@ -1,0 +1,3 @@
+export interface BlockSeparator {
+	id?: string;
+}
