+++ site/types/schema/content/page/block/index.ts
@@ -1,0 +1,9 @@
+export type * from './block-columns.js';
+export type * from './block-hero-form.js';
+export type * from './block-hero-headline.js';
+export type * from './block-hero-rotator.js';
+export type * from './block-logo-cloud.js';
+export type * from './block-media-full-width.js';
+export type * from './block-page-header.js';
+export type * from './block-separator.js';
+export type * from './block-showcase.js';
