+++ site/types/schema/content/page/component/component-button-group.ts
@@ -1,0 +1,6 @@
+import type { ComponentButton } from './component-button.js';
+
+export interface ComponentButtonGroup {
+	id?: string;
+	buttons?: (string | ComponentButton)[];
+}
