+++ site/types/schema/content/page/component/component-heading.ts
@@ -1,0 +1,6 @@
+export interface ComponentHeading {
+	id?: string;
+	preheading?: string | null;
+	heading?: string | null;
+	subheading?: string | null;
+}
