+++ site/types/schema/content/page/component/component-media.ts
@@ -1,0 +1,9 @@
+import type { Video } from '../../video.js';
+
+export interface CompMedia {
+	id?: string;
+	type?: string | null;
+	embed?: string | null;
+	video?: string | Video | null;
+	image?: string | File | null;
+}
