+++ site/types/schema/content/page/component/component-metrics.ts
@@ -1,0 +1,4 @@
+export interface ComponentMetrics {
+	id?: string;
+	data?: { [key: string]: any } | null;
+}
