+++ site/types/schema/content/page/component/index.ts
@@ -1,0 +1,2 @@
+export type * from './component-button-group.js';
+export type * from './component-button.js';
