+++ site/types/schema/content/page/index.ts
@@ -1,0 +1,4 @@
+export * from './block/index.js';
+export * from './component/index.js';
+export * from './page-section.js';
+export * from './page.js';
