+++ site/types/schema/content/report.ts
@@ -1,0 +1,3 @@
+export interface Report {
+	id?: string;
+}
