+++ site/types/schema/content/user-group.ts
@@ -1,0 +1,8 @@
+import type { File } from '../system/index.js';
+
+export interface UserGroup {
+	id?: number;
+	name?: string | null;
+	cover?: string | File | null;
+	link_url?: string | null;
+}
