+++ site/types/schema/content/video.ts
@@ -1,0 +1,8 @@
+export interface Video {
+	id?: string;
+	transcript?: string | null;
+	url?: string | null;
+	thumbnail?: string | File | null;
+	type?: string | null;
+	file?: string | File | null;
+}
