+++ site/types/schema/meta/index.ts
@@ -1,0 +1,4 @@
+export type * from './globals.js';
+export type * from './navigation.js';
+export type * from './redirect.js';
+export type * from './seo.js';
