+++ site/types/schema/routes/index.ts
@@ -1,0 +1,4 @@
+export * from './content-type.js';
+export * from './event.js';
+export * from './partner.js';
+export * from './team.js';
