+++ site/types/schema/routes/partner.ts
@@ -1,0 +1,11 @@
+import type { User } from '../system/index.js';
+
+export interface Partner {
+	id?: string;
+	status?: string;
+	sort?: number | null;
+	user_created?: string | User | null;
+	date_created?: string | null;
+	user_updated?: string | User | null;
+	date_updated?: string | null;
+}
