+++ site/types/schema/system/index.ts
@@ -1,0 +1,4 @@
+export type * from './file.js';
+export type * from './folder.js';
+export type * from './role.js';
+export type * from './user.js';
